PROTON-2559: making object.h mostly private
