+++ c/tools/reactor-recv.c
@@ -33,6 +33,7 @@
 #include "proton/url.h"
 #include "msgr-common.h"
 
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
