+++ ruby/cproton.i
@@ -22,6 +22,7 @@
 #define PN_USE_DEPRECATED_API 1
 
 #include <proton/connection_driver.h>
+#include <proton/cid.h>
 #include <proton/engine.h>
 #include <proton/handlers.h>
 #include <proton/message.h>
